+++ scripts/helpers.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 import time
 import contextlib
+import os
 
 class Timing(contextlib.ContextDecorator):
     def __init__(self, prefix="", on_exit=None, enabled=True): self.prefix, self.on_exit, self.enabled = prefix, on_exit, enabled
@@ -8,3 +9,9 @@ class Timing(contextlib.ContextDecorator):
     def __exit__(self, *exc):
         self.et = time.perf_counter_ns() - self.st
         if self.enabled: print(f"{self.prefix}{self.et*1e-6:6.2f} ms"+(self.on_exit(self.et) if self.on_exit else ""))
+
+def getenv(key):
+    var = os.getenv(key)
+    if not var or int(var) == 0:
+        return False
+    return True
